moved today screen functions to data object
