+++ app/src/main/java/co/develhope/meteoapp/today/TodayItem.kt
@@ -39,71 +39,3 @@ enum class TodayWeatherIcon(val image: Int){
     SUN_CLOUD(R.drawable.sun_cloud),
     MOON(R.drawable.moon)
 }
-
-fun createRandomValues(): MutableList<TodayData> {
-    val currentHour = OffsetTime.now().hour
-
-    val todayList = mutableListOf<TodayData>()
-
-    todayList.add(TodayData.TodayTitle("Palermo, Sicilia"))
-
-
-
-    var hour = currentHour
-
-    while(hour <= 23){
-        val image = R.drawable.sun
-        val degrees = (0..35).random()
-        val rainChance = (0..100).random()
-        val perceived = degrees - ((0..5).random())
-        val humidity = (0..100).random()
-        val wind = (0..5).random()
-        val coverage = (0..100).random()
-        val rainHeight = (0..5).random()
-
-        val todayData = createWeatherData(
-            hour,
-            image,
-            degrees,
-            rainChance,
-            perceived,
-            humidity,
-            wind,
-            coverage,
-            rainHeight
-        )
-        todayList.add(todayData)
-
-        hour++
-    }
-
-    return todayList
-
-}
-
-
-
-fun createWeatherData(
-    hour: Int,
-    image: Int,
-    degrees: Int,
-    rainChance: Int,
-    perceived: Int,
-    humidity: Int,
-    wind: Int,
-    coverage: Int,
-    rainHeight: Int
-): TodayData.TodayItem {
-    val time = OffsetTime.of(LocalTime.of(hour, 0), ZoneOffset.UTC)
-    val formattedTime = time.format(DateTimeFormatter.ofPattern("HH:mm"))
-    return TodayData.TodayItem(
-        formattedTime,
-        image,
-        degrees,
-        rainChance,
-        perceived,
-        humidity,
-        wind,
-        coverage,
-        rainHeight)
-}
